Cleans code per review.
