+++ src/JuliusSweetland.OptiKey.Core/Models/XmlKeyboardModels/XmlDynamicItem.cs
@@ -1,5 +1,4 @@
 ﻿// Copyright (c) 2019 OPTIKEY LTD (UK company number 11854839) - All Rights Reserved
-using JuliusSweetland.OptiKey.Enums;
 using System.Collections.Generic;
 using System.Xml.Serialization;
