+++ src/JuliusSweetland.OptiKey.Core/Models/XmlKeyboardModels/XmlKeyboard.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2019 OPTIKEY LTD (UK company number 11854839) - All Rights Reserved
 using log4net;
-using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using System.Windows;
@@ -158,4 +158,27 @@ namespace JuliusSweetland.OptiKey.Models
             return keyboard;
         }
     }
+
+    public class XmlVisualItem
+    {
+        protected static readonly ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
+
+        public XmlVisualItem() { }
+
+        [XmlElement("DynamicKey", typeof(XmlDynamicItem))]
+        [XmlElement("Scratchpad", typeof(XmlScratchpadItem))]
+        [XmlElement("SuggestionRow", typeof(XmlSuggestionRowItem))]
+        [XmlElement("SuggestionCol", typeof(XmlSuggestionColItem))]
+        public List<XmlVisualItem> Items { get; } = new List<XmlVisualItem>();
+
+        public int Row { get; set; } = -1;
+        public int Col { get; set; } = -1;
+        public int Width { get; set; } = 1;
+        public int Height { get; set; } = 1;
+        public string BackgroundColor { get; set; }
+    }
+
+    public class XmlScratchpadItem : XmlVisualItem { }
+    public class XmlSuggestionRowItem : XmlVisualItem { }
+    public class XmlSuggestionColItem : XmlVisualItem { }
 }
