+++ src/JuliusSweetland.OptiKey.Core/Models/XmlKeyboardModels/XmlSuggestionRow.xaml.cs
@@ -1,12 +1,7 @@
 ﻿// Copyright (c) 2019 OPTIKEY LTD (UK company number 11854839) - All Rights Reserved
 
-using System;
-using System.Globalization;
 using System.Windows;
 using System.Windows.Controls;
-using JuliusSweetland.OptiKey.Enums;
-using JuliusSweetland.OptiKey.Extensions;
-using JuliusSweetland.OptiKey.Properties;
 
 namespace JuliusSweetland.OptiKey.Models
 {
@@ -20,9 +15,6 @@ namespace JuliusSweetland.OptiKey.Models
             InitializeComponent();
             Loaded += (sender, args) => NumberOfSuggestionsDisplayed = 4;
         }
-
-        public int Row { get; set; }
-        public int Col { get; set; }
 
         public static readonly DependencyProperty NumberOfSuggestionsDisplayedProperty =
             DependencyProperty.Register("NumberOfSuggestionsDisplayed", typeof(int), typeof(XmlSuggestionRow), new PropertyMetadata(default(int)));
